API Refactor: carto.auth
- Moving Credential under auth folder/namespace
- Adding version to carto namespace
