+++ src/lib/auth/index.ts
@@ -1,0 +1,3 @@
+import { Credentials, setDefaultCredentials, defaultCredentials } from './Credentials';
+
+export { Credentials, setDefaultCredentials, defaultCredentials };
